修改：     apps/web-frontend/Makefile 	新文件：   apps/web-frontend/certs/server.crt 	新文件：   apps/web-frontend/certs/server.key 	新文件：   apps/web-frontend/certs/stellarium.crt 	新文件：   apps/web-frontend/certs/stellarium.key 	修改：     apps/web-frontend/kill_run.sh 	新文件：   apps/web-frontend/nginx.conf 	修改：     apps/web-frontend/public/favicon.ico 	修改：     apps/web-frontend/public/index.html 	新文件：   apps/web-frontend/server.py 	修改：     apps/web-frontend/src/App.vue 	修改：     apps/web-frontend/src/components/Chart-Focus.vue 	修改：     apps/web-frontend/src/components/FocuserPanel.vue 	修改：     apps/web-frontend/src/components/MountControlPanel.vue 	修改：     apps/web-frontend/src/components/Settings-Dialog-Focuser.vue 	修改：     apps/web-frontend/src/components/data-credits-dialog.vue 	修改：     apps/web-frontend/src/components/date-time-picker.vue 	修改：     apps/web-frontend/src/components/gui.vue 	修改：     apps/web-frontend/src/components/location-dialog.vue 	修改：     apps/web-frontend/src/components/location-mgr.vue 	修改：     apps/web-frontend/src/components/toolbar.vue 	修改：     apps/web-frontend/src/components/view-settings-dialog.vue 	新文件：   apps/web-frontend/stunnel.conf 	修改：     apps/web-frontend/vue.config.js 	新文件：   stunnel.conf

+++ apps/web-frontend/kill_run.sh
@@ -1,10 +1,21 @@
 # This tools is to kill the server , recompile the vue and c codes and restart the stellarium web server
 
+# Kill process on port 8080
 lsof -i:8080
 pid=$(sudo lsof -i:8080 | awk '{ if (NR!=1) { print $2 } }')
+if [ ! -z "$pid" ]; then
+    kill -9 $pid
+fi
+
+# Kill process on port 9090
+lsof -i:9090
+pid=$(sudo lsof -i:9090 | awk '{ if (NR!=1) { print $2 } }')
+if [ ! -z "$pid" ]; then
+    kill -9 $pid
+fi
+
 sed -i '/VUE_APP_VERSION/d' .env
 echo "VUE_APP_VERSION=$(date +%Y%m%d)" >> .env
-kill -9 $pid
 make update-engine
 make build
-make start
+make start-all
